+++ ext_map_case.c
@@ -6,7 +6,7 @@
 /*   By: bkotwica <bkotwica@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/30 09:50:22 by bkotwica          #+#    #+#             */
-/*   Updated: 2024/07/30 14:13:30 by bkotwica         ###   ########.fr       */
+/*   Updated: 2024/07/31 11:20:53 by bkotwica         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,8 @@
 
 int	check_top_right(char **map, int i, int j)
 {
-	if (i - 1 >= 0 && ft_strlen(map[i]) > j + 1 && ft_strlen(map[i - 1]) > j)
+	if (i - 1 >= 0 && (int)ft_strlen(map[i]) > j + 1
+		&& (int)ft_strlen(map[i - 1]) > j)
 	{
 		if ((map[i - 1][j] == '0' && map[i][j + 1] == '0') ||
 			(map[i - 1][j + 1] == '0' && map[i - 1][j] != '0'
@@ -26,8 +27,8 @@ int	check_top_right(char **map, int i, int j)
 
 int	check_bottom_right(char **map, int i, int j)
 {
-	if (list_len(map) > i + 1 && ft_strlen(map[i + 1]) > j
-		&& ft_strlen(map[i]) > j + 1)
+	if (list_len(map) > i + 1 && (int)ft_strlen(map[i + 1]) > j
+		&& (int)ft_strlen(map[i]) > j + 1)
 	{
 		if ((map[i + 1][j] == '0' && map[i][j + 1] == '0') ||
 			(map[i + 1][j + 1] == '0' && map[i][j + 1] != '0'
@@ -39,7 +40,8 @@ int	check_bottom_right(char **map, int i, int j)
 
 int	check_bottom_left(char **map, int i, int j)
 {
-	if (list_len(map) > i + 1 && j - 1 > 0 && ft_strlen(map[i + 1]) > j)
+	if (list_len(map) > i + 1 && j - 1 > 0
+		&& (int)ft_strlen(map[i + 1]) > j)
 	{
 		if ((map[i + 1][j] == '0' && map[i][j - 1] == '0') ||
 			(map[i + 1][j - 1] == '0' && map[i][j - 1] != '0'
@@ -51,7 +53,8 @@ int	check_bottom_left(char **map, int i, int j)
 
 int	check_top_left(char **map, int i, int j)
 {
-	if (j - 1 >= 0 && i - 1 >= 0 && ft_strlen(map[i - 1]) > j)
+	if (j - 1 >= 0 && i - 1 >= 0
+		&& (int)ft_strlen(map[i - 1]) > j)
 	{
 		if ((map[i - 1][j] == '0' && map[i][j - 1] == '0') ||
 			(map[i - 1][j - 1] == '0' && map[i - 1][j] != '0'
